# db: add BenchmarkIteratorScan